ML-75 Remove unused imports and refactor
