--- src/api/viewsets/__init__.py
@@ -1,2 +1,0 @@
-from .project.model import Project
-from .dataset.model import Dataset
